Migrated deck handling to gokerpeer, next step is to create middleman for GUI and gokerpeer
