+++ main.go
@@ -7,5 +7,3 @@ import (
 func main() {
 	gui.Init()
 }
-
-
